+++ firmware/ctrl/inc/firmware_app.h
@@ -2,5 +2,6 @@
 #define ACCEPTANCE_FIRMWARE_APP_H
 
 void App_Init(void);
+void App_StartUart(void);
 
 #endif